% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetrics.R
\name{cd}
\alias{cd}
\title{Comembership Distance}
\usage{
cd(c1, c2)
}
\arguments{
\item{c1}{A numeric vector of cluster labels}

\item{c2}{A numeric vector of cluster labels}
}
\value{
A numeric of the comembership distance between the two clusters
}
\description{
`cd` returns the comembership distance between two clusterings
of the same dataset. Each clustering must contain the same amount of clusters.
}
