% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distances.R
\name{compute_distances}
\alias{compute_distances}
\title{Compute Distance Metrics over K-Means clusterings}
\usage{
compute_distances(training, testing, Kmax = 6)
}
\arguments{
\item{training}{An nxp matrix of training data}

\item{testing}{An mxp matrix of testing data}

\item{Kmax}{The maximum number of clusters considered}
}
\value{
An (Kmax - 1) x 5 matrix of stability metrics.
Each row corresponds to a value of K. Each column corresponds to a particular metric.
}
\description{
`compute_distances` computes a set of stability metrics for
clusterings over two datasets, `training` and `testing`. For each number of clusters two through `Kmax`,
each dataset is clustered with `kmeans` and a set of stability distance metrics are computed over the
two clusterings. Each metric is designed such that the most stable clustering minimumises the metric.
The number of clusters `K` is determined from some selection rule applied over some aggregation of distance metrics
computed over some resampling method.
}
