% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distances.R
\name{getClusterAssignment}
\alias{getClusterAssignment}
\title{Get KMeans Cluster Assignment for new observations}
\usage{
getClusterAssignment(X, M, K)
}
\description{
`getClusterAssignment` computes the cluster assignments for a
 matrix of new observations, given a set of Kmeans centroids
 @param X data matix for which to compute cluster assignments
 @param M Kmeans centroids matrix
 @param K integer for number of centroids.
 @return vector of cluster indicies, assuming Kmeans clustering utlizes 1:K as cluster labels.
}
