% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetrics.R
\name{mmd}
\alias{mmd}
\title{Minimum Matching Distance}
\usage{
mmd(c1, c2, K)
}
\arguments{
\item{c1}{A numeric vector of cluster labels}

\item{c2}{A numeric vector of cluster labels}

\item{K}{Integer for number of clusters}
}
\value{
A numeric of the minimum matching distance between the two clusters
}
\description{
`mmd` returns the comembership distance between two clusterings
of the same dataset. Each clustering must contain the same amount of clusters.
}
