% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.R
\name{null_resample_bootstrap}
\alias{null_resample_bootstrap}
\title{Compute stability distance metrics through bootstrapping over null clustering distribution}
\usage{
null_resample_bootstrap(df, Kmax, nreps = 20)
}
\description{
`null_resample_boot` operated the same as `resample_boot`, but evaluates Kmeans
over samples fit from a uniform distribution over the principle components of
the original dataset`df`, following the null distribution from the Gap Statistic.
For each sample taken from the null distribution, the sample is split into a bootstrapped
sample and out-of-bag sample matching the computations done in `resample_boot`.

 @param `df` input dataset matrix
 @param nreps Number of samples taken from the null distribution
 @param Kmax Maximum number of clusters considered.
 @param nerps number of null samples taken
 @export
}
