% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.R
\name{null_resample_cv}
\alias{null_resample_cv}
\title{Compute stability distance metrics through CV over null clustering distribution}
\usage{
null_resample_cv(df, Kmax, Kfolds, nreps = 20)
}
\arguments{
\item{Kmax}{Integer for maximum number of clusters considered}

\item{Kfolds}{Number of folds used in cross-validation}

\item{nreps}{number of null samples taken}

\item{`df`}{input dataset matrix}
}
\description{
`null_resample_cv` operated the same as `resample_cv`, but evaluates Kmeans
over samples fit from a uniform distribution over the principle components of
the original dataset`df`, following the null distribution from the Gap Statistic.
For each sample taken from the null distribution, the sample is cut in proportions
 matching that of that of K-fold cross validation used to compute the stability
 metrics of the raw dataset.
}
