% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.R
\name{resample_boot}
\alias{resample_boot}
\title{Compute stability distance metrics through bootstrapping}
\usage{
resample_boot(df, B = 25, Kmax = 6)
}
\arguments{
\item{df}{A data matrix}

\item{B}{Integer for number of bootstrap sampled. Defaulted to 25.}

\item{Kmax}{Maximum number of clusters considered. Defaulted to 6.}
}
\description{
`resample_boot` utilizes bootstrapping to emulate resampling
}
