% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.R
\name{resample_cv}
\alias{resample_cv}
\title{Compute stability distance metrics through cross validation}
\usage{
resample_cv(df, K_folds = 4, Kmax = 6)
}
\arguments{
\item{df}{A data matrix}

\item{K_folds}{Integer for number of folds used in cross-validation. Defaulted to four.}

\item{Kmax}{Maximum number of clusters considered. Defaulted to 6.}
}
\description{
`resample_cv` utilized cross-validation to emulate resampling
}
