% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_analysis.R
\name{stability_analysis}
\alias{stability_analysis}
\title{Stability Analysis for Kmeans Clustering}
\usage{
stability_analysis(
  df,
  Kmax = 6,
  method = "CV",
  nreps = 4,
  null_dist = F,
  nreps_null = 30
)
}
\arguments{
\item{df}{Data matrix to be clustered. Each row is an observation, each column is a variable.}

\item{Kmax}{Integer for maximum number of clusters used by Kmeans. Larger values slow down computation. Defaulted to 6.}

\item{method}{String that is either "CV" or "boot" to select cross-validation or bootstrap resampling.
Defaulted to "CV", as bootstrapping often takes a larger number of samples to yield adequate performance.}

\item{nreps}{Integer for number of resamples to be taken.
If method is 'CV', `nreps` is the number of folds.
If method is 'boot', `nreps` is the number of bootstrap resamples taken.}

\item{null_dist}{Boolean for whether the stability metrics should be compared to metrics taken from
a data generating process with no clusters. Defaulted to 'False', as this is very computationally intensive.}

\item{nreps_null}{Integer for number of null samples taken in `null_dist` is set to True. Defaulted to 30 but more are recommended.}
}
\value{
An object of class `Kmeans_stability_analysis` which contains the mean stability metrics computed over `df`,
and a list of the observed stabiltiy metrics across each resample.
If `null_dist` is set to true, the mean stability for the null samples, the associated gao statistics, and the raw stability metrics for each null sample will also be returned.
Otherwise, those components of the `Kmeans_stability_analysis` object will be set to `NA`.
}
\description{
`stability_analysis` computes a set of stability metrics to estimate the most stable Kmeans clustering
Resampling approximations include K-fold cross validation and bootstrapping.
A clustering attaining the smallest instability is deemed the most stable clustering.
}
\details{
`stability_analysis` will compute five stability metrics for Kmeans clusterings for K ranging from two to `Kmax`.
Two metrics are cannon in the Kmeans stability literature: co-membership distance, and minimum matching distance.
The three other metrics are inspired by optimal transport distance. They are
(1) `wd` the minimal matching cost of the wasserstein distances for datapoints within two clusterings,
(2) `wd_cb` the Wasserstein distance for Kmeans centroids where each center is given uniform mass, and
(3) `wd_cp` ,the Wasserstein distance for Kmeans centroids where each center is given mass proportional to the number of observations within the associated cluster.

For all of these metrics, the number of clusters which attains the smallest value of each of these metrics is the clustering deemed most stable.
These metrics are computed through the the aid of the packages `transport` and `RcppHungarian`.

If `null_dist` is set to true,the five distance metrics are computed across clusterings of samples taken from a uniform distribution aligned to the priciple components of `df`.
Following the Gap Statistic, these valuesoperate as baseline values for the metrics under the setting of no clustering.
One may select the clustering which attains the largest difference in observed stability metric and the expected null metric to be the most stable clustering.
}
