% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_datagen.R
\name{toy_datagen}
\alias{toy_datagen}
\title{Toy Data Generating Process for Three Clusters}
\usage{
toy_datagen(ns = c(100, 60, 40))
}
\arguments{
\item{ns}{A numeric vector of length three. Each component is the number
of samples in one of the three clusters. Defaulted to c(100, 60, 40).}
}
\value{
A `sum(ns)` x 2 matrix.
}
\description{
`toy_datagen` constructs a two-dimensional data as a testing dataset.
}
