% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetrics.R
\name{wd}
\alias{wd}
\title{Minimum Wasserstein Distance}
\usage{
wd(K, training, testing, ctrain, ctest)
}
\arguments{
\item{K}{an integer for of clusters within both provided clusterings}

\item{training}{A matrix containing the data which generated the `ctrain` clustering vector}

\item{testing}{A matrix containing the data which generated the `ctest` clustering vector}

\item{ctrain}{A numeric vector containing the clustering of the `training` dataset}

\item{ctest}{A numeric vector containing the clustering of the `testing` dataset}
}
\value{
A numeric of the minimum wasserstein distance.
}
\description{
`wd` computes the minimum matching distance between two clusterings
of two distinct datasets. The optimal transport distance between points within
each pair of clusters is computed, from which the optimal matching of clusters is computed.
}
