% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceMetrics.R
\name{wd_cp}
\alias{wd_cp}
\title{Minimum Wasserstein Distance over Cluster Centers with Proportional Mass}
\usage{
wd_cp(K, cluster_training, cluster_testing)
}
\arguments{
\item{K}{Integer for number of clusters}

\item{cluster_training}{A `kmeans` return object built from clustering
a training dataset with `K` clusters
#' @param cluster_testing A `kmeans` return object built from clustering
a testing dataset with `K` clusters}
}
\value{
A numeric of the wasserstein distance between the centroids of the two clusterings
}
\description{
`Wd_cp` computes the wasserstein distance between K_means cluster
centroids. The cluster centroids are given mass proportional to the number of
}
