% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp_modified.R
\name{wpp_modified}
\alias{wpp_modified}
\title{A Modified Version of `wpp`.}
\usage{
wpp_modified(coordinates, mass)
}
\arguments{
\item{coordinates}{A data matrix.}

\item{mass}{A vector of masses.}
}
\value{
A `wpp` object.
}
\description{
`wpp_modified` is a stripped down version of `wpp` from the `transport` package
designed to allow construction of `wpp.object` objects which are a single point mass.
The current implementation of `wpp` prohibits this, but this is necessary for
computing `wd` when a single point is assigned its own cluster. See `wpp` in the `transport` package for more details.
}
